Applications/Research_Agent/pyproject.toml
[project]
name = "ai-research-assistant"
version = "1.0.0"
description = "An AI assistant that assists in research"
authors = [
    {name = "Virendrasinh Chavda"}
]
readme = "README.md"

[tool.setuptools.packages.find]
where = ["src"]

requires-python = ">=3.12, <3.14"
dependencies = [
]


[tool.poetry.dependencies]
python = ">=3.12, <3.14"

# Core dependencies
python-dotenv = "^1.0.1"
pydantic = "^2.12.3"
pydantic-settings = "^2.1.0"

# Data processing
arxiv = "^2.3.0"
tqdm = "^4.67.1"
pymupdf = "^1.26.6"
pandas = "^2.3.3"

# NLP and embeddings
sentence-transformers = "^5.1.2"
transformers = "^4.40.0"

# Vector database
qdrant-client = "^1.15.1"

# Database
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"  # PostgreSQL adapter

# Cache
redis = "^5.0.0"

# Web framework
django = "^5.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
django-channels = "^4.0.0"
channels-redis = "^4.1.0"

# LLM providers
openai = "^1.6.0"
anthropic = "^0.20.0"
google-generativeai = "^0.4.0"

# Async
aiohttp = "^3.9.0"
asyncio-contextmanager = "^1.0.0"

# LangGraph and LangChain
langgraph = "^0.0.50"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-community = "^0.0.30"
langchain-openai = "^0.0.10"
langchain-anthropic = "^0.1.0"

# Development
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
